Add TODOs in Timer class and optimize app initialization
An optimization has been made in 'app.js' to initiate the function `init()` earlier for better execution flow. In 'timer.js', TODO comments have been inserted to indicate required future changes such as removing static implementation of the Timer and to understand if the class singleton is the default export.